+++ apps/service/src/features/auth/hooks/use-auth.ts
@@ -1,7 +1,7 @@
 import { handleError } from "@/app/error-handler/error-handler";
 import { useLogin } from "@/features/auth/hooks/use-login";
 import { useLogout } from "@/features/auth/hooks/use-logout";
-import { useGetUserInfo } from "@/pages/my/hooks/useGetUserInfo";
+import { useGetUserInfo } from "@/pages/my/hooks/use-get-user-info";
 import { ROUTE_PATH } from "@app/routes";
 import { BadRequestError } from "@repo/api";
 import { toast } from "@repo/ui";
