+++ apps/service/src/pages/available-classes/available-classes.page.tsx
@@ -1,6 +1,6 @@
 import { AvailableClasses } from "@/pages/available-classes/components/available-classes";
 import { ClassesFilterSelect } from "@/pages/available-classes/components/classes-filter-select";
-import { useAvailableClassesFilter } from "@/pages/available-classes/hooks/useAvailableClassesFilter";
+import { useAvailableClassesFilter } from "@/pages/available-classes/hooks/use-available-classes-filter";
 import { MenteeTheme } from "@/shared/components/mentee-theme/mentee-theme";
 import { PageLayout } from "@/shared/layouts/page-layout";
 import { AspectRatio, Typography } from "@repo/ui";
