+++ apps/service/src/pages/available-classes/hooks/use-get-available-classes.ts
