+++ apps/service/src/pages/available-classes/hooks/use-get-class-detail.ts
