+++ apps/service/src/pages/available-classes/hooks/use-get-job-categories.ts
